Moved websocket logic outside of zustand
